Fixed bug with multipeer
